+++ src/Laravel/Filter/WhereBelongsTo.php
@@ -2,7 +2,6 @@
 
 namespace Tobyz\JsonApiServer\Laravel\Filter;
 
-use Illuminate\Contracts\Database\Query\Expression;
 use Tobyz\JsonApiServer\Context;
 use Tobyz\JsonApiServer\Schema\Type\Arr;
 use Tobyz\JsonApiServer\Schema\Type\Str;
@@ -34,21 +33,24 @@ class WhereBelongsTo extends Where
     protected function applyValue(object $query, mixed $value, Context $context): void
     {
         if ($this->asBoolean) {
-            $query->{$value ? 'whereNotNull' : 'whereNull'}($this->getColumn($query));
+            [$column, $bindings] = $this->getColumn($query, $context);
+
+            $this->addColumnBindings($query, $bindings);
+            $query->{$value ? 'whereNotNull' : 'whereNull'}($column);
             return;
         }
 
         parent::applyValue($query, $value, $context);
     }
 
-    protected function getColumn(object $query): string|Expression
+    protected function getColumn(object $query, Context $context): array
     {
-        if ($this->column) {
-            return $this->column;
+        if ($this->column !== null) {
+            return parent::getColumn($query, $context);
         }
 
         $relationship = $query->getModel()->{$this->relationship ?: $this->name}();
 
-        return $relationship->getQualifiedForeignKeyName();
+        return [$relationship->getQualifiedForeignKeyName(), []];
     }
 }